+++ mayan/apps/user_management/apps.py
@@ -2,6 +2,7 @@ from __future__ import unicode_literals
 
 from django.apps import apps
 from django.contrib.auth import get_user_model
+from django.contrib.auth.signals import user_logged_in, user_logged_out
 from django.db.models.signals import post_save
 from django.utils import six
 from django.utils.translation import ugettext_lazy as _
@@ -28,7 +29,10 @@ from .events import (
     event_group_created, event_group_edited, event_user_created,
     event_user_edited
 )
-from .handlers import handler_initialize_new_user_options
+from .handlers import (
+    handler_initialize_new_user_options, handler_user_logged_in,
+    handler_user_logged_out
+)
 from .links import (
     link_current_user_details, link_current_user_edit, link_group_create,
     link_group_delete, link_group_edit, link_group_list, link_group_user_list,
@@ -261,5 +265,17 @@ class UserManagementApp(MayanAppConfig):
             receiver=handler_initialize_new_user_options,
             sender=User
         )
+
+        user_logged_in.connect(
+            dispatch_uid='user_management_handler_user_logged_in',
+            receiver=handler_user_logged_in,
+            sender=User
+        )
+        user_logged_out.connect(
+            dispatch_uid='user_management_handler_user_logged_out',
+            receiver=handler_user_logged_out,
+            sender=User
+        )
+
         registry.register(Group)
         registry.register(User)