+++ mayan/apps/user_management/events.py
@@ -21,3 +21,9 @@ event_user_created = namespace.add_event_type(
 event_user_edited = namespace.add_event_type(
     label=_('User edited'), name='user_edited'
 )
+event_user_logged_in = namespace.add_event_type(
+    label=_('User logged in'), name='user_logged_in'
+)
+event_user_logged_out = namespace.add_event_type(
+    label=_('User logged out'), name='user_logged_out'
+)