+++ HISTORY.rst
@@ -206,6 +206,7 @@
 * Add outline links CSS for facets.
 * Add a bottom margin to list links.
 * Use copyfileobj to save documents to files
+* Add user logged in and logged out events.
 
 3.1.11 (2019-04-XX)
 ===================